Projeto-coracao-peludo/backend/requirements.txt
fastapi
uvicorn
motor
python-jose
passlib[bcrypt]
pydantic
email-validator
python-dotenv